+++ figma/src/flutter/make/text.make.ts
@@ -48,26 +48,19 @@ export function makeText(node: ReflectTextNode): Text {
 
 
 function escapeDartString(text: string): string {
-    const splittedChars = text.split("\n");
-    text = splittedChars.length > 1 ? splittedChars.join("\\n") : text;
-
-
     // \ -> \\
     text = text.split("\\").join("\\\\");
 
+    // \n -> \\n
+    text = text.split("\n").join("\\n");
+
     // $ -> \$''"
-    // const re_dollar_sign = new RegExp('\$', 'g');
     text = text.split('$').join('\\$');
-    // text = text.replace(re_dollar_sign, '\\$')
 
     // " -> \"
-    // const re_double_quote = new RegExp('\"', 'g');
-    // text = text.replace(re_double_quote, '\\"')
     text = text.split('"').join('\\"');
 
     // ' -> \'
-    // const re_single_quote = new RegExp("\'", 'g');
-    // text = text.replace(re_single_quote, "\\'")
     text = text.split("'").join("\\'");
 
     return text;